Move BudgetTransaction into its own module
